+++ CHANGES.md
@@ -23,6 +23,7 @@
 - remove compression framing support
 - remove tls-auth support
 - remove "multi instance" support
+- update to version 4 of `fkooman/oauth2-server`
 
 ## 1.8.5 (2018-11-28)
 - no longer allow clients to obtain new access tokens using the refresh token